Parse XLSX files with unusual structure
Normally you will find `workbook.xml` in the `xl` directory, but recently we got a file where it was located in the root.

To find the actual location of workbook, we must parse the file `_rels/.rels`. A separate rels file located relative to the workbook then points out the locations of the sheets, styles and shared strings.